requirements.txt
flask==2.2.2
Flask_SQLAlchemy==3.0.3
flask-mysql==1.5.2
Flask-MySQLdb==1.0.1
python_dotenv==0.21.1
flask-marshmallow==0.15.0
PyJWT==2.8.0